+++ odoo/addons/test_orm/security/ir.access.csv
@@ -221,3 +221,5 @@ access_test_orm_domain_expression_users,test_orm_domain_expression_users,test_or
 access_test_orm_business,test_orm_business,test_orm.business,base.group_user,crud,
 access_test_orm_groups,test_orm_groups,test_orm.groups,base.group_user,crud,
 access_test_orm_properties,test_orm_properties,test_orm.properties,base.group_system,crud,
+access_test_orm_inherits_partner,test_orm_inherits_partner,test_orm.inherits.partner,base.group_user,crud,
+access_test_orm_inherits_users,test_orm_inherits_users,test_orm.inherits.users,base.group_user,crud,
